refactor(api): extract INJECT/KILL/DELETE/SPAWN to handlers/instance.rs
Step 3 Slice B — follows HandlerCtx pattern from Slice A.

New file: src/api/handlers/instance.rs
- handle_inject(), handle_kill(), handle_delete(), handle_spawn()

handle_session changes:
- INJECT/KILL/DELETE/SPAWN arms → one-liner calls
- Remaining 3 arms: CREATE_TEAM, UPDATE_TEAM, SHUTDOWN

## Refactor Pattern Change
All 4 arms converted from writeln(writer)+continue to return Value.
Error responses are now returned as Value and written by the dispatch
loop's normal response path. Behavioral equivalence: same JSON content
written to the same stream, one line per request.

## Disclosed Risk
SPAWN happy path requires real agent binary + PTY — cannot be
characterized. SPAWN dedup (already-exists check) requires a
pre-populated managed registry which needs a real spawn — also
cannot be characterized. Both are known gaps.

All other error paths (missing name, validate_name, backend not found,
working_directory rejected) are covered by Commit 1.

Post-merge: manual SPAWN smoke check to verify InstanceCreated event
and log correctness.

Ref: Slice A Pattern Established (b8d5c40)

Tests: 615 pass, 0 fail, 2 ignored.
