+++ src/api/handlers/mod.rs
@@ -4,6 +4,7 @@
 //! `HandlerCtx` bundles the session-scoped state that handlers need.
 
 pub(crate) mod external;
+pub(crate) mod instance;
 pub(crate) mod messaging;
 pub(crate) mod query;
