+++ src/api/mod.rs
@@ -309,215 +309,10 @@ fn handle_session(
 
         let response = match method {
             method::LIST => handlers::query::handle_list(params, &ctx),
-            method::INJECT => {
-                let name = params["name"].as_str().unwrap_or("");
-                if let Err(e) = agent::validate_name(name) {
-                    let _ = writeln!(writer, "{}", json!({"ok": false, "error": e}));
-                    continue;
-                }
-                let data = params["data"].as_str().unwrap_or("");
-                // "raw" flag: send bytes as-is (for attach-like paths)
-                let raw = params["raw"].as_bool().unwrap_or(false);
-                let reg = agent::lock_registry(registry);
-                match reg.get(name) {
-                    Some(handle) => {
-                        // Check if agent is restarting
-                        let is_restarting = handle
-                            .core
-                            .lock()
-                            .map(|c| c.state.current.is_unavailable())
-                            .unwrap_or(false);
-                        if is_restarting {
-                            json!({"ok": false, "error": format!("agent '{name}' is restarting, retry later")})
-                        } else {
-                            let result = if raw {
-                                agent::write_to_agent(handle, data.as_bytes())
-                            } else {
-                                agent::inject_to_agent(handle, data.as_bytes())
-                            };
-                            match result {
-                                Ok(()) => json!({"ok": true, "result": {"bytes": data.len()}}),
-                                Err(e) => json!({"ok": false, "error": format!("{e}")}),
-                            }
-                        }
-                    }
-                    None => {
-                        let ext = agent::lock_external(externals);
-                        if ext.contains_key(name) {
-                            json!({"ok": false, "error": format!("agent '{name}' is external — use send instead of inject")})
-                        } else {
-                            json!({"ok": false, "error": format!("agent '{name}' not found")})
-                        }
-                    }
-                }
-            }
-            method::KILL => {
-                let name = params["name"].as_str().unwrap_or("");
-                if let Err(e) = agent::validate_name(name) {
-                    let _ = writeln!(writer, "{}", json!({"ok": false, "error": e}));
-                    continue;
-                }
-                let reg = agent::lock_registry(registry);
-                match reg.get(name) {
-                    Some(handle) => {
-                        if let Ok(mut core) = handle.core.lock() {
-                            core.state.set_restarting();
-                        }
-                        let mut child = crate::sync::lock_poisoned(&handle.child, "api_child");
-                        let _ = child.kill();
-                        drop(child);
-                        drop(reg);
-                        crate::event_log::log(home, "kill", name, "killed via API");
-                        json!({"ok": true})
-                    }
-                    None => {
-                        // Try external registry
-                        drop(reg);
-                        let mut ext = agent::lock_external(externals);
-                        if ext.remove(name).is_some() {
-                            crate::event_log::log(home, "kill", name, "external agent removed");
-                            json!({"ok": true})
-                        } else {
-                            json!({"ok": false, "error": format!("agent '{name}' not found")})
-                        }
-                    }
-                }
-            }
-            method::DELETE => {
-                // Kill + remove from registry + configs
-                let name = params["name"].as_str().unwrap_or("");
-                if let Err(e) = agent::validate_name(name) {
-                    let _ = writeln!(writer, "{}", json!({"ok": false, "error": e}));
-                    continue;
-                }
-                // Check external registry first
-                {
-                    let mut ext = agent::lock_external(externals);
-                    if ext.remove(name).is_some() {
-                        crate::event_log::log(home, "delete", name, "external agent deleted");
-                        let _ = writeln!(writer, "{}", json!({"ok": true}));
-                        let _ = writer.flush();
-                        continue;
-                    }
-                }
-                // Kill and remove from registry first — this prevents the PTY close
-                // handler from sending a crash event (agent not in registry = no crash).
-                let mut reg = agent::lock_registry(registry);
-                if let Some(handle) = reg.get(name) {
-                    let mut child = crate::sync::lock_poisoned(&handle.child, "api_child");
-                    let _ = child.kill();
-                    drop(child);
-                }
-                reg.remove(name);
-                drop(reg);
-                // Then remove config (no race: agent already gone from registry)
-                crate::sync::lock_poisoned(configs, "api_configs").remove(name);
-                // Cleanup the agent's published port file
-                crate::ipc::remove_port(&crate::daemon::run_dir(home), name);
-                crate::event_log::log(home, "delete", name, "deleted via API");
-                // Notify TUI to close the corresponding pane
-                if let Some(n) = notifier {
-                    tracing::info!(agent = name, "DELETE emitting InstanceDeleted");
-                    n.notify(ApiEvent::InstanceDeleted {
-                        name: name.to_string(),
-                    });
-                }
-                json!({"ok": true})
-            }
-            method::SPAWN => {
-                let name = match params["name"].as_str() {
-                    Some(n) => n,
-                    None => {
-                        let _ =
-                            writeln!(writer, "{}", json!({"ok": false, "error": "missing name"}));
-                        continue;
-                    }
-                };
-                if let Err(e) = agent::validate_name(name) {
-                    let _ = writeln!(writer, "{}", json!({"ok": false, "error": e}));
-                    continue;
-                }
-                // Dedup: reject spawn for an already-registered name. spawn_agent
-                // silently overwrites the registry entry, which would orphan the
-                // previous agent's PTY and leave panes reading from a stale
-                // subscription. The respawn-after-crash path in daemon.rs goes
-                // through a different code path and is not affected.
-                if agent::lock_registry(registry).contains_key(name) {
-                    let _ = writeln!(
-                        writer,
-                        "{}",
-                        json!({"ok": false, "error": format!("agent '{name}' already exists")})
-                    );
-                    continue;
-                }
-                let command = params["backend"]
-                    .as_str()
-                    .map(String::from)
-                    .unwrap_or_else(|| {
-                        crate::fleet::FleetConfig::load(&home.join("fleet.yaml"))
-                            .ok()
-                            .and_then(|f| {
-                                f.defaults.backend.map(|b| b.preset().command.to_string())
-                            })
-                            .unwrap_or_else(|| "claude".to_string())
-                    });
-                let command = command.as_str();
-                let args: Vec<String> = params["args"]
-                    .as_str()
-                    .map(|s| s.split_whitespace().map(String::from).collect())
-                    .unwrap_or_default();
-                let requested_work_dir = params["working_directory"]
-                    .as_str()
-                    .map(std::path::PathBuf::from)
-                    .unwrap_or_else(|| home.join("workspace").join(name));
-                let work_dir = match validate_working_directory(&requested_work_dir, home) {
-                    Ok(p) => p,
-                    Err(e) => {
-                        let _ =
-                            writeln!(writer, "{}", json!({"ok": false, "error": format!("{e}")}));
-                        continue;
-                    }
-                };
-                let size = crossterm::terminal::size().unwrap_or((120, 40));
-                let spawn_mode = match params["mode"].as_str() {
-                    Some("resume") => crate::backend::SpawnMode::Resume,
-                    _ => crate::backend::SpawnMode::Fresh,
-                };
-
-                match spawn_one(
-                    home, registry, name, command, &args, spawn_mode, &work_dir, size,
-                ) {
-                    Ok(()) => {
-                        if let Some(n) = notifier {
-                            let layout_hint =
-                                LayoutHint::parse(params["layout"].as_str().unwrap_or("tab"));
-                            let spawner = params["spawner"]
-                                .as_str()
-                                .filter(|s| !s.is_empty())
-                                .map(String::from);
-                            let target_pane = params["target_pane"]
-                                .as_str()
-                                .filter(|s| !s.is_empty())
-                                .map(String::from);
-                            tracing::info!(
-                                agent = name,
-                                layout = ?layout_hint,
-                                spawner = ?spawner,
-                                target_pane = ?target_pane,
-                                "SPAWN emitting InstanceCreated"
-                            );
-                            n.notify(ApiEvent::InstanceCreated {
-                                name: name.to_string(),
-                                layout: layout_hint,
-                                spawner,
-                                target_pane,
-                            });
-                        }
-                        json!({"ok": true, "result": {"name": name}})
-                    }
-                    Err(e) => json!({"ok": false, "error": format!("{e}")}),
-                }
-            }
+            method::INJECT => handlers::instance::handle_inject(params, &ctx),
+            method::KILL => handlers::instance::handle_kill(params, &ctx),
+            method::DELETE => handlers::instance::handle_delete(params, &ctx),
+            method::SPAWN => handlers::instance::handle_spawn(params, &ctx),
             method::SEND => handlers::messaging::handle_send(params, &ctx),
             method::STATUS => handlers::query::handle_status(params, &ctx),
             method::REGISTER_EXTERNAL => handlers::external::handle_register_external(params, &ctx),
